+++ test-projects/social-login/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone/cypress-project-facebook-login
+
 # @keystone-alpha/cypress-project-facebook-login
 
 ## 2.2.9