+++ test-projects/social-login/CHANGELOG.md
@@ -1,4 +1,40 @@
+# @keystone/cypress-project-facebook-login
+
+## 5.0.0
+
+### Major Changes
+
+- [`c7ba40ec`](https://github.com/keystonejs/keystone-5/commit/c7ba40ec98116603c6b7a501d2442e16170ec6be) [#1813](https://github.com/keystonejs/keystone-5/pull/1813) Thanks [@jesstelford](https://github.com/jesstelford)! - - This is the first release of `@keystone/*` packages (previously `@keystone-alpha/*`).
+  - All packages in the `@keystone-alpha` namespace are now available in the `@keystone` namespace, starting at version `5.0.0`.
+  - To upgrade your project you must update any `@keystone-alpha/*` dependencies in `package.json` to point to `"@keystone/*": "^5.0.0"` and update any `require`/`import` statements in your code.
+
+### Patch Changes
+
+- Updated dependencies [[`c7ba40ec`](https://github.com/keystonejs/keystone-5/commit/c7ba40ec98116603c6b7a501d2442e16170ec6be)]:
+  - @keystone/adapter-mongoose@5.0.0
+  - @keystone/app-admin-ui@5.0.0
+  - @keystone/app-graphql@5.0.0
+  - @keystone/app-static@5.0.0
+  - @keystone/auth-passport@5.0.0
+  - @keystone/auth-password@5.0.0
+  - @keystone/fields@5.0.0
+  - @keystone/file-adapters@5.0.0
+  - @keystone/keystone@5.0.0
+
 # @keystone-alpha/cypress-project-facebook-login
+
+## 2.2.9
+
+### Patch Changes
+
+- Updated dependencies [[`0a36b0f4`](https://github.com/keystonejs/keystone-5/commit/0a36b0f403da73a76106b5e14940a789466b4f94), [`7129c887`](https://github.com/keystonejs/keystone-5/commit/7129c8878a825d961f2772be497dcd5bd6b2b697), [`3bc02545`](https://github.com/keystonejs/keystone-5/commit/3bc025452fb8e6e69790bdbee032ddfdeeb7dabb), [`768420f5`](https://github.com/keystonejs/keystone-5/commit/768420f567c244d57a4e2a3aaafe628ea9813d9d), [`a48281ba`](https://github.com/keystonejs/keystone-5/commit/a48281ba605bf5bebc89fcbb36d3e69c17182eec), [`a8ee0179`](https://github.com/keystonejs/keystone-5/commit/a8ee0179842f790dd3b5d4aae3524793e752ee26), [`effc1f63`](https://github.com/keystonejs/keystone-5/commit/effc1f639d5824720b7a9d82c2ee881d77acb901)]:
+  - @keystone-alpha/keystone@16.1.0
+  - @keystone-alpha/app-graphql@8.2.1
+  - @keystone-alpha/adapter-mongoose@6.0.1
+  - @keystone-alpha/app-admin-ui@5.10.3
+  - @keystone-alpha/fields@15.0.0
+  - @keystone-alpha/auth-passport@4.1.5
+  - @keystone-alpha/auth-password@1.0.6
 
 ## 2.2.8